requirements.txt
aio_pika
aiohttp
django
django-ninja
pika
python-dotenv